chore: updates for Rust 1.66

+++ syncserver/src/db/mysql/models.rs
@@ -403,7 +403,7 @@ impl MysqlDb {
                 collection: bso.collection.clone(),
                 collection_id,
             })?;
-            if usage.total_bytes >= self.quota.size as usize {
+            if usage.total_bytes >= self.quota.size {
                 let mut tags = HashMap::default();
                 tags.insert("collection".to_owned(), bso.collection.clone());
                 self.metrics.incr_with_tags("storage.quota.at_limit", tags);
@@ -489,7 +489,7 @@ impl MysqlDb {
                 bso::expiry,
             ))
             .filter(bso::user_id.eq(user_id))
-            .filter(bso::collection_id.eq(collection_id as i32)) // XXX:
+            .filter(bso::collection_id.eq(collection_id)) // XXX:
             .filter(bso::expiry.gt(now))
             .into_boxed();
 
@@ -572,7 +572,7 @@ impl MysqlDb {
         let mut query = bso::table
             .select(bso::id)
             .filter(bso::user_id.eq(user_id))
-            .filter(bso::collection_id.eq(collection_id as i32)) // XXX:
+            .filter(bso::collection_id.eq(collection_id)) // XXX:
             .filter(bso::expiry.gt(self.timestamp().as_i64()))
             .into_boxed();
