+++ syncserver/src/db/tests/db.rs
@@ -625,7 +625,7 @@ async fn get_collection_usage() -> Result<()> {
             db.put_bso(pbso(
                 uid,
                 coll,
-                &format!("b{}", i as i32),
+                &format!("b{}", i),
                 Some(&String::from_utf8_lossy(&payload)),
                 None,
                 None,
